#65: factor out material stream getter from arnold material node
