+++ src/serlio/prtMaterial/MaterialUtils.h
@@ -5,10 +5,13 @@
 #include "maya/MPlug.h"
 #include "maya/MStatus.h"
 #include "maya/MString.h"
+#include "maya/adskDataStream.h"
 
 #include <map>
 
 namespace MaterialUtils {
+
+adsk::Data::Stream* getMaterialStream(MObject& aOutMesh, MObject& aInMesh, MDataBlock& data);
 
 MStatus getMeshName(MString& meshName, const MPlug& plug);
